Add DB naming conventions and migration hygiene test
